file size returns long long.
git-svn-id: file:///fltk/svn/trunk@5776 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/filename.h
@@ -96,7 +96,7 @@ inline char* filename_ext(char* a) {return (char*)(filename_ext((const char*)a))
 FL_API bool filename_match(const char *, const char *pattern); // glob match
 FL_API bool filename_exist(const char*);
 FL_API bool filename_isdir(const char*);
-FL_API double filename_size(const char *); // return size of file
+FL_API long long unsigned filename_size(const char *); // return size of file
 FL_API long int filename_mtime(const char *); // return modification time
 
 typedef int (File_Sort_F)(const dirent*const*, const dirent*const*);
